+++ integration_tests/pages/apply/risks-and-needs/health-needs/communicationAndLanguagePage.ts
@@ -1,7 +1,7 @@
-import { Cas2Application as Application } from '../../../../server/@types/shared/models/Cas2Application'
-import ApplyPage from '../applyPage'
-import paths from '../../../../server/paths/apply'
-import { pageIsActiveInNavigation } from './utils'
+import { Cas2Application as Application } from '../../../../../server/@types/shared/models/Cas2Application'
+import ApplyPage from '../../applyPage'
+import paths from '../../../../../server/paths/apply'
+import { pageIsActiveInNavigation } from '../utils'
 
 export default class CommunicationAndLanguagePage extends ApplyPage {
   constructor(private readonly application: Application) {
